feat: Adds client features endpoint
This is worth a comment, since the way it was solved depends on which
mode you're executing edge in. Currently edge only supports one mode, so
we could've just explicitly passed in the provider for getting features.

However, we are aware that we want at least two more providers for
features, so this lays the ground work for implementing more modes and
providers, but still using the same actix handler for the
`/api/client/features` endpoint.

Co-authored-by: Simon Hornby <liquidwicked64@gmail.com>