feat: lock free feature resolution
Redesign the way data flows through Edge. Previously, we had thread locks on our data sources, which was impacting the response time of the application. This moves everything to be in memory cached with lazy persistence in the background and reloading the state on application startup. This means the hot path is now lock free.

---------

Co-authored-by: sighphyre <liquidwicked64@gmail.com>