fix: Move token cache resolution out of FromRequest
This was unfortunatley the wrong place to have it, it meant we never hit
the middleware which validated our token. So unfortunately, we had to
make our endpoints depend on the tokencache in order to have the up to
date EdgeToken, rather than the one that comes in with the request.

I fear this has broken Offline Mode, so not quite ready to merge just
yet