fix: Use split_once. (#164)
We had forgotten that `:` is a valid part of a header value, so
splitting on `'` ended up splitting into too many parts and thus failing
the parser. This PR changes to use split_once, which splits on first
occurrence. Since Header Names are not allowed to contains `:` this will
be fine