Testing if organising features in ENV -> PROJECT -> FEATURES is easier to work with, compared to ENV -> FEATURES.
I think it is, but it needs more work to be efficient. I think the old
way of storing should be removed in favour of this way. The way of
returning toggles should be changed to select by project and environment
instead of selecting by environment and then filtering by projects.