initial draft of refreshing projects
Currently we only upsert, that is, we keep all data that exists in our
cached list, but replace toggles with the same key with the updated
toggle configuration as well as add new toggles.

However, if a toggle is removed from upstream and the new refresh comes
down with this toggle missing, it will still be present in the edge
cache after a refresh has happened.

What this PR does (poorly) is make sure we only keep toggles from
projects that is not contained in the updated list of toggles coming in.

I want some more tests for various behaviours, but all current tests
around this logic now runs green, so I haven't broken any behaviour we
already depended on.

Definitely open to optimisations. I felt filthy writing something that
is O(3n). At least it isn't O(n^2)