fix: No longer return wrong feature toggle
Previously, upon first access with a new token which Edge hadn't seen
before our rehydration code only filtered by project of the token, so
once it had the data it would stupidly return the token that was
available for the new token (in the wrong project).

After this refactor/fix we now apply all filters to all queries for
feature toggles from the cache, which should reduce the combinations of
possible filter mistakes.

fixes: #276