fix: Wildcard token refreshing.
Previously our refreshing algorithm assumed that we always had at least
one project with an explicit name for deciding what to keep. When in
reality we could have a wildcard token (*) for updating. This means that
we should just return the update as our new set of data for this token.

Added another test with a wildcard token to verify that we do indeed
only keep the update.

In addition, added a filter for unique feature names to what we're
keeping to avoid mixing multiple features of the same name.