fix: Only replace projects covered by token
Previously, if the update was empty, we assumed we needed to replace all
stored features. This is not the case. We only need to remove the
projects that are connected to the token we're using for the merge.

In addition, debug output on `/internal-backstage/tokens` now includes how many features was received in
the update