feat: add debian 12 build file
Also made the dockerignore file just ignore the target folder except for
the binaries. Which will allow us to have a docker image that does the
building.

The debian12 recipe uses cargo chef to have better layer caching, as
well as debian:bookworm-slim as the base image for runtime, which is
slightly bigger than distroless.

However, it does not build for both amd64 and aarch yet