fix: bootstrap reload in offline mode. (#595)
After adding support for client and frontend tokens, we did not extend
the reloader to check the client and frontend token arrays. This PR extends
tokens with FE and Client tokens, to ensure that we refresh the data for
all our tokens.

In addition we make /internal-backstage/tokens useful for offline mode
as well, to at least be able to see which tokens you added to Edge.

Fixes: #594