feat: streaming now enabled/disabled with runtime flag (#605)
This PR takes streaming from being a compile-time feature, to a feature controlled via the CLI. The new arg `streaming` (requires `strict`) will cause Edge to start up in streaming mode instead of polling.

Note that streaming updates from Unleash is still experimental and this is not stable. But this'll allow us to start testing it in our own infrastructure.