feat: Enable consumption metrics by default (#1081)
Removes the flag for consumption reporting, enabling it everyone all of the time.

In theory, this might warrant a minor version bump, but there was a docstring that said that changes to this feature _may_ not be follow semver (which is about as useless as it gets), so I'm guessing ... we can make it a patch? Eh, we needn't decide on that now, anyway

~~Still need to figure out whether this messes with Edge when you don't run strict.~~ This makes it so that consumption is run even in dynamic mode. From what I understand, this shouldn't be an issue at all. A connection is still a connection. @kwasniew also said that the `requires strict` was just a copy-paste from a different flag, so that seems safe.