feat(1-3968): send edge hosting type (self-hosted/hosted) with instance metrics data (#1086)
Adds edge hosting type to the edge instance data schema. The hosting type can be set via the env var `EDGE_HOSTING`. 

In serializing, we serialize it as either `hosted` or `self-hosted`. Edge will parse any value that isn't `hosted` (regardless of capitalization) as `self-hosted`. If you don't provide a value, it will not set that value in the instance data payload.

By making the property optional, we also let Edge accept incoming edge data schemas that don't have the property set.

Adds basic tests to verify that we send the property if and only if the value is set.