fix: Normalize path now works (#1253)
* fix: Normalize path now works

With the axum upgrade we did remember to add NormalizePath layer, but
we added it as a post router layer, which meant that the router ran
before the path was normalizaed and therefore never used the normalized
path for routing. This PR updates to running NormalizePath before the
entire router, which again makes us support "/api/frontend" and
"/api/frontend/".

* chore(nit): removed unnecessary comment

* chore: added integration test for path normalization

* fix: path update after subfolder move of crates

* chore(rustfmt): formatting